Clean up removal of all non-AEAD modes
The draft was in a sort of halfway state between whether TLSCiphertext
has a fragment that is a selection or just embedding the AEAD elements
directly.  This should normalize it to the embedded view without the
extra GenericAEADCipher struct definition.

(i don't think this changes the logic of the draft at all)