Move ContentType inside encryption
Previous versions of TLS have left the ContentType in the clear
without justifying the exposure of this piece of information.

This change moves the real ContentType inside the encryption to
protect its confidentiality.  It leaves the external ContentType, but
fixes its value to application_data(23) to appease silly middleboxes.

For clarity, this also means redefining the NULL_NULL case to not
simply be an pseudo-cipher that implements the identity
transformation, but to explicitly state it as sending the TLSPlaintext
struct directly over the wire.  ekr has agreed to re-write that text.