limit CertificateRequest only to position in the handshake
I don't think the CertificateRequest needs to be limited to
authenticated servers, or to certain ciphersuites.

i note that RFC 5246 (TLS 1.2) does seem to suggest these same
limitations, but it's not clear to me why they exist.  If we're going
for a simpler state machine for TLS 1.3, i see no reason to keep these
limitations.