specify ClientHello record layer version
Freeze the ClientHello record layer version to { 3, 1 } as per suggestion from Brian Smith.
http://www.ietf.org/mail-archive/web/tls/current/msg14870.html

Logically, the initial ClientHello is TLS 1.0 as it is compatible with TLS 1.0 clients. Setting the initial version to reflect this makes sense. This replaces the previous vague list of possible recommendations in the backwards compatibility section. This is in line with practices currently employed by some existing TLS implementations.