Copy edit certificate_list changes
Remove self-signed, since the key point is that it's a trust anchor that was independently distributed.  Note here that some trust anchors are signed, though that usually indicates a reason for it to be included, the signing might persist for longer than the actual need for it.

Also the explanation had more words than it needed.

n.b., I only loosely wrapped this stuff, if you want a proper wrapping of lines, name a preferred wrap column, because I couldn't work out what to do here.