Implement padding in ciphertexts by a block zeros after content type.
At the TLS WG Interim meeting in Redmond, September 2015 we had a
lengthy discusion about possible content padding schemes.  We decided
provisionally on this scheme, which injects padding as a string of
zero octets at the end of the cleartext sent to the AEAD algorithm.

The buffer of zeros is delimited from the plaintext content by one
non-zero octet of ContentType.

Other schemes considered had the ContentType at the end of the
clearted, with an explicit padding length field placed just before it.

If we had chosen that approach, we would need to decide how to encode
the explicit padding length field, and how to avoid sending an extra
padding byte if no padding was desired by the sender.

The proposed scheme is marginally better than these other schemes
because:

 * no new content_type value needs to be allocated
 * arbitrary padding size is possible (0 up to max record size)
 * peers enforce that padding must be all-zeros (no new heartbleed)
 * any content type can be padded in the same way