Add a section on 0-RTT backwards compatibility.
0-RTT is *not* backwards compatible because of the external application data
records. Add some guidance for both clients and servers to mitigate this.
Ideally the server SHOULD could be a MUST and that would be the end of it, but
I do not expect all servers to get this right and so clients will want a
fallback.

This also lets us provide guidance on the exact client-side fallback trigger
and explicitly say not to change the version. Conveniently, there is a reliable
signal available. (Historically, clients have used network errors or generic
protocol errors as fallback triggers which masks new server bugs.)