Add some more implementation pitfalls.
There are many fields where unknown values are supposed to be ignored and not a
fatal error. Experience with buggy TLS 1.2 servers suggests this is somehow not
obvious. Switch the mention of omitting the extensions field completely (which
is no longer possible) with unknown extensions.

Also explicitly call out ClientHello version processing as this historically
has been a huge ecosystem failure.