Remove redundant length prefix from KeyShareEntry.key_exchange contents.
Only the outer length prefix is necessary, not the inner one. Named groups can
just as easily take a byte string with known length rather than need to parse
fields out.

This is consistent with the definition of DigitallySigned.signature; the
signature is an opaque field which is passed as a byte string into the
corresponding signature algorithm. The definition of rsa_pss_sha256 does not
define an inner u16 length prefix. It simply takes the opaque byte string and
passes it into the rsa_pss_sha256 function.

(In fact, this change means KeyShareEntry and DigitallySigned structures
exactly parallel.)

This also fixes some references to non-existent structures like ECPoint and
dh_Yc. (Those seem to be remnants of TLS 1.2.)