Define ecdsa_sha1 and mark it as a SHOULD NOT alongside other SHA-1 algorithms.
The revised signature algorithm mechanism left the {sha1, ecdsa}
SignatureAndHashAlgorithm without a corresponding SignatureScheme defined at
all. SHA-1 does not have a corresponding curve, so defining it alongside
ecdsa_p256_sha256 and friends was a nuisance.

But omitting it, combined with other text, implicitly means it "MUST NOT be
offered or negotiated by any implementation".  Clients (ideally) send only one
ClientHello, so this implies a constraint on TLS 1.2 as well. SHA-1 needs to
go, but I think I was a tad overzealous here.

I wouldn't be surprised {sha1, ecdsa} can be removed from TLS 1.2 right now.
Certainly it will be much much easier to remove than {sha1, rsa}. But this is
probably better a SHOULD NOT rather than a MUST NOT. We wouldn't like to
discover after the fact that a MUST NOT in TLS 1.3 is untenable due to some
ecosystem problem. Notably, old versions of some popular server software ignore
signature algorithms and only sign SHA-1. This mostly affects RSA since ECDSA
isn't that common, but any such ECDSA deployments would be affected.

This is also a little odd for servers as TLS 1.2 says missing sigalgs means the
client is assumed to support {sha1, X}. (Although, strictly speaking, the spec
does not say the server must do anything useful with that advertisement.)

In contrast, dsa_sha1 is already not offered by many major browsers but is
merely a SHOULD NOT.

Define ecdsa_sha1_RESERVED as a companion to dsa_*_RESERVED to bring it out of
the implicit MUST NOT. Then include it in the SHOULD NOT lists alongside
rsa_pkcs1_sha1 and dsa_sha1.