Use more capital letters to describe version negotiation.
People will continue getting this wrong until the end of time, but add an extra
MUST in the server_version half on the off chance someone reads it.

Also clarify that client_version is a maximum version. ("The version of the TLS
protocol by which the client wishes to communicate during this session"
suggests it should be an exact match.)