Define the interaction between 0-RTT and multiple PSK identities.
Unlike both TLS 1.2 resumption and TLS 1.2 PSK, TLS 1.3 increases the state
space and allows clients to offer multiple PSK identities. This mostly works,
but 0-RTT presumes that only one is offered. Define 0-RTT to presume the first
identity is selected and clarify that other shapes are possible due to the
increased state space.