Negotiating 0-RTT means that the server has 1.3 everywhere.
The text that @davidben proposed that covered 0-RTT and backward
compatibility took the position that a server might not support
TLS 1.3 when it was attempted.  That leads to all sorts of mess,
particularly given that most 1.2 and earlier servers are totally
incapable of handling 0-RTT messages.  In NSS, it's a fatal
error and I would expect others to be similarly strict given the
smackTLS work.

This is a simpler set of rules:
 * If the server advertises support for 0-RTT that means that it
   commits to supporting TLS 1.3 for the duration.
 * If the client attempts 0-RTT, then it should abort if it sees
   a ServerHello with 1.2 or earlier.

This is stricter than what we currently have, but it's crisper.
That is, unless people imagine modifying their TLS 1.2
implementations to be able to ignore 0-RTT data,