Move ticket_age to the early_data extension
This allows us to remove EncryptedExtensions from the 0-RTT flight.
Doing this means that the server has all the information that they
need to generate a ServerHello in the ClientHello.  That reduces
the complexity of the state machine needed considerably.

The 0-RTT Finished message is only needed for key confirmation for
0-RTT, which means that processing the Finished is not necessary.
If it's wrong, then the connection can fail.  That's different to
ticket_age, which would be "wrong" very often, so that we could
only reject 0-RTT.

This reduces generality somewhat, but since we've not identified
any other use for EncryptedExtensions, this isn't the end of the
world.  The other candidate use for EncryptedExtensions was
encrypted SNI, but it's not clear that a design based on
EncryptedExtensions is the right way to solve that problem.

A bad server can unmask clients by using a zero, fixed, or
predictable mask.  I don't believe that this is worth complicating
the design to counter.  This is a case where the dumbest possible
solution is the best.