Move the downgrade sentinel to the end of server_random.
The sentinel value was chosen to avoid conflicting with any valid TLS 1.2
value, but this is not actually desirable. We need to avoid conflicting with
any non-TLS-1.3 server but *still* be a valid TLS 1.2 value. A TLS 1.3 server
must still correctly implement TLS 1.2 to interoperate with existing TLS 1.2
clients.

In particular, placing it at the front clobbers TLS 1.2's server timestamp.
Putting this in the random was silly and 1.3 rightfully removes it, but some
software depends on servers sending a timestamp. The tlsdate program uses this
as a secure-ish time synchronization protocol. Silly as it is, it exists and it
arguably was following the letter of the TLS 1.2 spec, even if we don't like
that timestamp very much...

Move the sentinel to the end so servers may implement this downgrade protection
(which will be needed in the short term given the current version negotiation
as it forces browsers to implement the version fallback again) without breaking
any tlsdate instances which may still exist.