Remove server missing_extension requirements.
(This is orthogonal to whether alerts should be MUST or SHOULD. I've
left them as MUST for now.)

While the client can reasonably send missing_extension in response to a
ServerHello which, say, selected an ECDHE cipher without key_share, the
server is in no such position. Such a requirement is unhelpful as a
natural implementation would have the same behavior (connection failed)
anyway.

Since servers already must account for some ciphers requiring ECDHE and
some not, a server may implement logic such as:

    ecdheOk := tryToSelectAGroup(myList, clientList)
    for cipher in serverCiphers {
        if cipher not in clientCiphers {
            continue
        }
        if isECDHE(cipher) and not ecdheOk {
            continue
        }
        ... other checks, etc. ...
        yay, I found a cipher
    }
    if no cipher was found {
        send handshake_failure and close connection
    }

There is no reason to ask a server to distinguish between the following
cases, all of which will fail anyway:

- The client sent only ECDHE ciphers I know about but did not send
  key_share.
- The client sent only ECDHE ciphers I know about but only key_shares I
  don't implement.
- The client sent only ECDHE ciphers, some of which I don't know about,
  and did not send key_share.
- The client sent ECDHE ciphers I know about and some PSK ciphers I
  don't know about and did not send key_share.
- The client sent only ECDHE ciphers, none of which I know about, and
  did not send key_share.

Notably, the some cases are not even distinguishable, even though the
old text mandates different alerts for them. The only case we can
reasonably send missing_extension for is the first, but this is an
unhelpful special-case to mandate. It only adds complexity (the
connection will fail anyway), thus risking bugs.