Clarify that the key changes immediately at Finished.
As the Finished message is an implicit ChangeCipherSpec, it's worth
clarifying that this is the point at which the cipher state changes.

Should an implementation not send half-RTT data, it may defer switching
traffic keys until both Finished messages have been exchanged. This will
not work well, as alerts sent in response to a bad client certificate
will be unreadable. The outgoing traffic keys must be switched early.