Improve the guidance around supporting TLS 1.2.
The document was pretty crisp about the version downgrade protections, but
was less good about the changes to signature schemes.  I have made a few
changes to the description of signature_algorithms that should improve that.
In particular, I have made it clear that the RSASSA-PSS codepoints are also
defined for TLS 1.2 and that they can be used there.  That means that TLS
1.3 implementations need to be prepared to accept PSS signatures.

This also adds a section that points to the pieces that 1.2 implementations
need to consider.