Rename various version fields.
From the past three times we've bumped the version, reality has
unequivocally told us that the TLS version negotiation, as described in
previous specifications, has been a collossal failure.

The working group decided not to replace the mechanism with a more
sustainable one, so do the next best thing in trying to save TLS from
itself. If the fields are renamed to what they actually are
('server_version' is the true version and 'client_version' is a
maximum), then perhaps we'll have some hope of future implementors
getting it right, if we ever clear through the prior iterations of the
bug.

While I'm here, remove the mention of 'The protocol version the current
record is compatible with' for record_version. record_version is purely
vestigial.