Forbid post-handshake handshake/application interleave.
No sender could possibly need to send half of a message interleaved with
application data in TLS, so forbid this. This cuts down on the
state-space a receiver needs to worry about.

This also means one cannot force the peer to buffer all but one byte of
a giant post-handshake message, and then never complete it for the
lifetime of the connection. (Although it does not remove all
steady state buffering DoS-type concerns.)