Fis PSK-related types
The PskKeMode and PskAuthMode types were not defined, substitute them by
the defined enums. Use non-plural enum names for consistency with other
enumeration types.

Remove the old opaque psk_identity type in favor of PskIdentity which
was otherwise unused. Increase minimum length from 2 to 6 to account for
two non-empty fields and their length and the the PskIdentity.identity
field length.