Clarify the behavior of no exporter context.
No context and empty context are now declared to be the same. Add some
ALL CAPS WORDS for future specifications to deprecate no context. I used
SHOULD to forbid no context because, if some protocol currently uses no
context, switching it to empty context is a breaking change and probably
not worth it. We just want new ones to always provide a context.

This affects the rules for using exporters, so RFC 5705 is added to the
list of updated documents.

Closes issue #546.