Remove server_name from the 0-RTT section.
This seems to have just been a mistake? Unlike ALPN, server_name isn't a
value that is selected by the server, so there's no need to match it
against a value stored in the session.

There is the issue of whether sessions are resumable across names or
other application-level contexts (for instance, whether we send client
certificates, different cipher profiles, etc.), but that is a more
complex issue than simply matching on server_name on the server side.
Moreover, any rules there would want to apply to resumption as a whole.
It seems there's no reason to have 0-RTT-specific rules over resumption
ones here.