Add certificate_required alert.
There is no alert for a server which requested a client certificate,
required it, but was not sent one. Servers today send a generic
handshake_failure which means clients must use heuristics to show error
messages to the user.

Unlike nearly every other negotiation in TLS, client certificate
selection is actually sometimes a user-level decision, so it is arguably
the most important place for us to have a dedicated alert.

(Also remove a "This alert is always fatal" note on
bad_certificate_status_response. All alerts are now fatal, so this is
redundant.)