Renumber new SignatureSchemes.
Avoid both starting with an existing hash AND ending with an existing
SignatureAlgorithm. In principle, this should not be necessary, but TLS
1.2 says:

   signature
      This field indicates the signature algorithm that may be used.
      The values indicate anonymous signatures, RSASSA-PKCS1-v1_5
      [PKCS1] and DSA [DSS], and ECDSA [ECDSA], respectively.  The
      "anonymous" value is meaningless in this context but used in
      Section 7.4.3.  It MUST NOT appear in this extension.

Bouncy Castle appears to have been enforcing this, causing interop
issues. As this was a straight-up MUST NOT, it seems prudent to
renumber. And if we're renumbering anyway, may as well avoid all the old
SignatureAlgorithms and avoid confusing tools like Wireshark.