Tighten and clarify KeyShareEntry ordering.
The text already says that both supported_groups and key_shares are in
order of "client preference". Supposing "client preference" is a
well-defined notion, we are already requiring the order match.

The SHOULD was also unclear on whether the list may be a general
subsequence. Clarify that this is allowed. This may be useful if, say,
we define some very large NamedGroup (post-quantum?) in the future that
is still preferable over the existing options. Clients would likely wish
to NOT offer it initially because it would be a waste of bandwidth for
the vast majority of servers which don't support it. (Not to mention
intolerance problems.)

In that scenario, a client would likely not predict a key_share by
default and instead use some kind of server info cache (either caching
what the server chose or using the full server supported_groups list).
But that client would still wish to advertise the new group first.