Editorial
Problems with non-compliant servers caused us to change the
version negotiation scheme, so we should hedge our "interoperably
negoatiate" statement a bit.

Note that the extensions called out in the handshake diagrams
are not intended to be comprehensive.

Move discussion of clients receiving TLS 1.3 ServerHellos during
(forbidden) renegotiation to the ServerHello section; it does
not really seem in place for the ClientHello section.

Tidy up the ClientHello extension presentation language.  Extensions
are mandatory for TLS 1.3, and the shortest key_share seems allowed
by the presentation language seems to be 4 octets of KeyShareEntry,
two octets of length, and two octets of ExtensionType.  A (pre-TLS 1.3)
ClientHello with no extensions is not indicated by a zero-length vector,
but rather by the absense of the field entirely.  Use the appropriate
syntax for that, and adjust the corresponding text a bit.

Lots of generic grammar/language fixes and similar.