change language on retransmission on 0-RTT data
The previous language made it seem like if a TLS stack would choose to retransmit the data on 0-RTT rejection, it was doing the wrong thing. However TLS stacks should be free to choose how they want to deal with these cases. For example they could tear down the connection if ALPN changes. Practically these cases should almost never happen so these cases are very rare.

I think putting this restriction on TLS implementations is a bit unfair. As far as I can tell, cases where it's hard to retransmit 0-RTT data happen in cases where the abstraction boundary is violated when application features add TLS features as a performance optimization, for example in the case of ALPN and token binding. As with any performance optimization it's fair to ask for a slow path when the optimization fails.