Consolidated version of Kaduk's PR#810.
The commit messages are below.

Editorial

legacy_compression_methods is a fixed byte string (*)

(*) when sending.  Servers that can negotiate <1.3 must still
accept other values, but this is the 1.3 spec, so we don't need
to allow for that in the presentation language specification.

SH extensions cannot be empty

Editorial

Editorial

Editorial

TLSPlaintext.legacy_record_version is not constant

A TLS 1.3 server, at least, should not enforce these constant
bytes since it may receive a ClientHello from a different TLS
version.

TLSCiphertext legacy_record_version MAY be enforced

It is supposed to always be 0x0301, and endpoints can check that.

Editorial

Rework the field names in the Record Proection section

Many of these were missed when TLSInnerPlaintext was added
as a distinct type.

Revamp introduction to alerts

We seem to be strongly deprecating the alert severity field
and making the severity implicit in the type of the alert.
Rather than sticking to a text that reflects the TLS 1.2 state
of affairs, use fresher text that is a better match for the new
perspective.

Reorder backwards compat security restrictions a bit

Put the no-sslv2 and no-sslv3 restrictions together, while keeping
no-sslv2 and no-v2-compatible-ClientHello together as well.

Move the MUST NOTs relating to versions less than 0x0300 before
the SHOULD NOT.

Adapt to feedback from @davegarrett