Various ServerHello and HelloRetryRequest nitpicks.
This is a giant pile of nitpicks, loosely inspired by the FFDHE alert
discussion.

The old text specifies a handshake_failure which is redundant with and
contradicts the Cryptographic Negotiation section. The old text says
only handshake_failure is allowed while Cryptographic Negotiation allows
both handshake_failure and insufficient_security. Fix this by avoiding
redundancy as the failure cases are global to the overall negotiation
algorithm.

Next, align the tense with itself and the rest of the text. Most of the
text appears to be written in present tense. The old text for SH said:

  The server *will send* [...] when it *was* able to [...]

which is a little odd. Also align the wording between the two. Talk
about a server (singular) to match Cryptographic Negotiation and say
"parameters" rather than "algorithms" or "algorithms and groups".

Also be consistent about the SH vs. HRR condition. The SH section said
"the client’s key_share extension was acceptable" while the HRR section
said "the client’s ClientHello did not contain sufficient information to
proceed with the handshake". Use the more general one.