Added text on determining the validity of OCSP responses
The OCSP protocol (RFC6960), defines the validity period of an
OCSP response based on the presence of the thisUpdate and nextUpdate
values. However, the nextUpdate value is optional, and may be omitted
in an OCSP response. In that case it is implied by the OCSP issuer
that an updated OCSP response is available at all time, however,
that does not give a hint to verifier for how long to treat the
previous response valid. Set a maximum value to 15 days.