Updating second ClientHello based on HelloRetryRequest
Section 4.1.2 lists the specific items that may be changed when sending a second ClientHello in response to a HelloRetryRequest. Since the HelloRetryRequest specifies a server_version and cipher_suite, then it seems that there is no reason for the second ClientHello to offer any version or cipher suite other than the ones in the HelloRetryRequest.