Allow the padding extension to change on HRR.
The padding extension is typically computed as part of serializing the
ClientHello, in hopes of targetting a particular size. As specified right now,
the second ClientHello must not use this same logic and instead must retain the
previous extension sizethough this would likely not hit the same target size.

The server's going to ignore it anyway, so allow it to be recalcuated. This
avoids unnecessary state in clients and extra logic to serialize the first and
second ClientHellos differently.