Adjust the wording around KCI attacks
The previous wording implied that a session in which an actor's key has been compromised should enjoy "peer authentication", which is defined just above to mean that both parties should authenticate each other. This adjusts the wording to make it clear what we intend, which is that compromising Alice's key should not allow you to impersonate Bob to Alice.