Update draft-ietf-tls-rfc8446bis.md
As discussed on the TLS list. It seems like cipher suites without confidentiality is already registered. Feel free to reformulate in any way. Refering to the cipher suites in the document is not very useful for the reader, and it does not seem to be a rule that a NULL encryption could not be recommended even if that seems very unlikely.

I also suggest adding "(certificate)" after client identity. The security property does obviously not hold for PSK authentication.