Certificate-Based Client Authentication
Based on the PSK authentication was not included in the past. I think specification has some places where "Client Authentication" refer to only certificate-based client authentication and not client authentication is general.

Not a big thing, but I think it might confuse some readers, and would be good to clarify.