Reduce emphasis on timing in the 0-RTT cache attack
This attack is often possible even without a timing
channel due to application-layer behaviors that allow
cache probing (e.g. in HTTP and DNS).

This change addresses the original concern in #1225.
A more thorough revamp of the anti-replay and side channel
recommendations might be needed to address all the
questions raised there.