Fix reference to MGF1
MGF1 got unhelpfully spelled out as "mask generation function 1" in
RFC 8446. I'm guessing this is the result of an acronym expansion
editting pass, but the function is simply called "MGF1". There is no
such thing as "mask generation function 1".

Correct this back to MGF1. In hopes this doesn't get hit the same
editing pass later, I've tweaked the wording and moved the citation so
that "mask generation function" is still uttered before MGF1, and it's
clearer that both primitives are defined in RFC 8017.